+++ app/models/comment_vote.rb
@@ -4,7 +4,6 @@ class CommentVote < ApplicationRecord
 
   validates_presence_of :score
   validates_uniqueness_of :user_id, :scope => :comment_id, :message => "have already voted for this comment"
-  validate :validate_comment_can_be_down_voted
   validates_inclusion_of :score, :in => [-1, 1], :message => "must be 1 or -1"
 
   after_create :update_score_after_create
@@ -23,12 +22,6 @@ class CommentVote < ApplicationRecord
   def self.search(params)
     q = search_attributes(params, :id, :created_at, :updated_at, :score, :comment, :user)
     q.apply_default_order(params)
-  end
-
-  def validate_comment_can_be_down_voted
-    if is_positive? && comment.creator == CurrentUser.user
-      errors.add(:base, "You cannot upvote your own comments")
-    end
   end
 
   def is_positive?